compute接口函数
# message auto-generated for no-merge-commit merge:
merge compute into master

summary下接口函数修改

Created-by: insightDeveloper

Match-id-f1e68d79d699e5d231b73264a5f2e56d093fd00e

+++ server/src/modules/summary/protocol/SummaryProtocolResponse.h
@@ -74,7 +74,7 @@ struct ComputeDetail {
 struct ComputeDetailResponse : public Response {
     ComputeDetailResponse() : Response(REQ_RES_COMPUTE_DETAIL) {}
     std::vector<ComputeDetail> computeDetails;
-    int32_t totalNum;
+    int64_t totalNum;
 };
 } // end of namespace Protocol
 } // end of namespace Dic
